// src/views/withdraw/enums/beneficiary.enum.ts
export enum BeneficiaryType {
  INTERNATIONAL = 'INTERNATIONAL',
  DOMESTIC = 'DOMESTIC',
  ACH_PAD = 'ACH_PAD',
}

export enum CounterpartyType {
  CRYPTO = 'CRYPTO',
  FIAT = 'FIAT',
}

export enum CounterpartyStatus {
  REGISTERED = 'REGISTERED',
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  REJECTED = 'REJECTED',
}

export enum CounterpartyType {
  CRYPTO_BTC = 'CRYPTO',
  FIAT_US = 'FIAT_US',
}

export enum NetworkBank {
  WIRE = 'FEDWIRE',
  SWIFT = 'SWIFT',
  ACH = 'ACH',
}

export enum TypeBankDetails {
  INTERNATIONAL = 'INTERNATIONAL',
  NATIONAL = 'DOMESTIC',
}